Refactor DLP validation error UI into its own component
