+++ web/src/views/DandisetLandingView/DandisetPublish.vue
@@ -153,124 +153,7 @@
       </v-tooltip>
     </v-row>
 
-    <v-row
-      v-if="currentDandiset.status === 'Pending'"
-      class="my-2 px-1"
-      no-gutters
-    >
-      <v-menu
-        :nudge-width="200"
-      >
-        <template #activator="{ on: menu, attrs }">
-          <v-tooltip bottom>
-            <template #activator="{ on: tooltip }">
-              <v-card
-                class="amber lighten-5 no-text-transform"
-                outlined
-                v-bind="attrs"
-                v-on="{ ...tooltip, ...menu }"
-              >
-                <v-row class="align-center px-4">
-                  <v-col
-                    cols="1"
-                    class="justify-center py-0"
-                  >
-                    <v-icon
-                      color="warning"
-                      class="mr-1"
-                    >
-                      mdi-playlist-remove
-                    </v-icon>
-                  </v-col>
-                  <v-spacer />
-                  <v-col cols="9">
-                    <div
-                      v-if="currentDandiset"
-                      class="text-caption"
-                    >
-                      Validation of the dandiset is pending.
-                    </div>
-                  </v-col>
-                </v-row>
-              </v-card>
-            </template>
-            <span>Reload the page to see if validation is over.</span>
-          </v-tooltip>
-        </template>
-      </v-menu>
-    </v-row>
-
-    <!-- Dialog where version and asset errors are shown -->
-    <v-dialog v-model="errorDialogOpen">
-      <ValidationErrorDialog
-        :selected-tab="selectedTab"
-        :asset-validation-errors="currentDandiset.asset_validation_errors"
-        :version-validation-errors="currentDandiset.version_validation_errors"
-        :owner="isOwner"
-        @openMeditor="openMeditor"
-      />
-    </v-dialog>
-
-    <!-- Version Validation Errors Button -->
-    <v-card
-      v-if="currentDandiset.version_validation_errors.length"
-      class="my-2 px-1 amber lighten-5 no-text-transform"
-      outlined
-      @click="openErrorDialog('metadata')"
-    >
-      <v-row class="align-center px-4">
-        <v-col
-          cols="1"
-          class="justify-center py-0"
-        >
-          <v-icon
-            color="warning"
-            class="mr-1"
-          >
-            mdi-playlist-remove
-          </v-icon>
-        </v-col>
-        <v-spacer />
-        <v-col cols="9">
-          <div
-            v-if="currentDandiset"
-            class="text-caption"
-          >
-            This Dandiset has {{ currentDandiset.version_validation_errors.length }}
-            metadata validation error(s).
-          </div>
-        </v-col>
-      </v-row>
-    </v-card>
-
-    <!-- Asset Validation Errors Button -->
-    <v-card
-      v-if="numAssetValidationErrors"
-      class="my-2 px-1 amber lighten-5 no-text-transform"
-      outlined
-      @click="openErrorDialog('assets')"
-    >
-      <v-row class="align-center px-4">
-        <v-col
-          cols="1"
-          class="justify-center py-0"
-        >
-          <v-icon
-            color="warning"
-            class="mr-1"
-          >
-            mdi-database-remove
-          </v-icon>
-        </v-col>
-        <v-spacer />
-        <v-col cols="9">
-          <div class="text-caption">
-            This Dandiset has {{ numAssetValidationErrors }}
-            asset validation error(s).
-          </div>
-        </v-col>
-      </v-row>
-    </v-card>
+    <DandisetValidationErrors :dandiset="currentDandiset" :isOwner="isOwner" />
 
     <v-row>
       <v-subheader class="mb-2 black--text text-h5">
@@ -354,15 +237,14 @@ import moment from 'moment';
 
 import type { RawLocation } from 'vue-router';
 import { useRoute } from 'vue-router/composables';
+
 import { dandiRest, loggedIn as loggedInFunc, user } from '@/rest';
 import { useDandisetStore } from '@/stores/dandiset';
 import router from '@/router';
 import type { User, Version } from '@/types';
-
 import { draftVersion } from '@/utils/constants';
-import { open as meditorOpen } from '@/components/Meditor/state';
+import DandisetValidationErrors from './DandisetValidationErrors.vue';
 
-import ValidationErrorDialog from '@/components/DLP/ValidationErrorDialog.vue';
 
 const PUBLISH_CHECKLIST = [
   'A descriptive title (e.g., <span class="font-italic">Data related to foraging behavior in bees</span> rather than <span class="font-italic">Smith et al 2022</span>)',
@@ -490,30 +372,11 @@ onUnmounted(() => {
   window.clearInterval(timer);
 });
 
-// Error dialog
-const errorDialogOpen = ref(false);
-type ErrorCategory = 'metadata' | 'assets';
-const selectedTab = ref<ErrorCategory>('metadata');
-function openErrorDialog(tab: ErrorCategory) {
-  errorDialogOpen.value = true;
-  selectedTab.value = tab;
-}
 
-function openMeditor() {
-  errorDialogOpen.value = false;
-  meditorOpen.value = true;
-}
 
-const numAssetValidationErrors = computed(() => {
-  if (currentDandiset.value === null) {
-    return 0;
-  }
-
-  return currentDandiset.value.asset_validation_errors.length;
-});
 const publishButtonDisabled = computed(() => !!(
   currentDandiset.value?.version_validation_errors.length
-      || numAssetValidationErrors.value
+      || currentDandiset.value?.asset_validation_errors.length
       || currentDandiset.value?.dandiset.embargo_status !== 'OPEN'
       || publishDisabledMessage.value
 ));
